split of query building bits from searchhelper
